Reorganize test classes to prevent unittest from running the base TestCases
