+++ netbox/utilities/testing/__init__.py
@@ -1,0 +1,2 @@
+from .testcases import *
+from .utils import *
